+++ app/builders/pathfinder2_character/classes/rogue_builder.rb
@@ -3,10 +3,8 @@
 module Pathfinder2Character
   module Classes
     class RogueBuilder
-      # rubocop: disable Metrics/AbcSize
       def call(result:)
         result[:main_ability] = 'dex' unless result[:main_ability]
-        result[:health] = { current: result[:health] + 8, max: result[:health] + 8, temp: 0 }
         result[:abilities].merge!({ result[:main_ability].to_sym => 2 }) { |_, oldval, newval| oldval + newval }
         result[:skill_boosts].merge!({ stealth: 1, free: 7 }) { |_, oldval, newval| oldval + newval }
 
@@ -18,7 +16,6 @@ module Pathfinder2Character
 
         result
       end
-      # rubocop: enable Metrics/AbcSize
     end
   end
 end