+++ app/platforms/pathfinder2/character.rb
@@ -15,7 +15,6 @@ module Pathfinder2
     attribute :classes, array: true
     attribute :subclasses, array: true
     attribute :abilities, array: true, default: { str: 10, dex: 10, con: 10, int: 10, wis: 10, cha: 10 }
-    attribute :health, array: true
     attribute :languages, array: true, default: []
     attribute :selected_skills, array: true, default: {} # { 'acrobatics' => 2, 'arcana' => 1, 'crafting' => 3 }
     attribute :lores, array: true, default: {}
@@ -36,10 +35,14 @@ module Pathfinder2
     attribute :selected_features, array: true, default: {} # { 'fighting_style' => ['fighting_style_defense'] }
     attribute :spent_spell_slots, array: true, default: {}
     attribute :experience, :integer, default: 0
+    attribute :health_current, :integer, default: 6
+    attribute :health_temp, :integer, default: 0
     # только для 1 уровня
     attribute :ability_boosts, array: true # дополнительные повышения характеристик
     attribute :ability_boosts_v2, array: true, default: {} # дополнительные повышения характеристик
     attribute :skill_boosts, array: true, default: {} # дополнительные повышения навыков
+    # DEPRECATED
+    attribute :health, array: true
   end
 
   class Character < Character